Use BaseDirs for locating font directories
Identifying the correct platform-specific font directories can be done
mostly correctly most of the time just with a hardcoded list of places
to look (particularly on Apple systems). However, on Windows and
XDG-following (i.e. Linux and friends) systems the user and system font
directories can end up in other places, and accommodating these edge
cases takes some effort.

This effort has already been put in with the BaseDirs package, which is
a small zero-dependency package whose entire purpose is to find the
correct directories for different types of content on various platforms.
So, we can remove the current font-folder-finding code entirely and just
call BaseDirs.fonts() for a reduction in code here and an improvement in
compatibility.