Prioritise certain file extensions in findfont
It's entirely possible that multiple forms of a font may be present on
one system, for instance an OTF and also a PFB from texlive. We care
about this because different formats have different levels of
capability, and we don't want to choose a worse one simply because it's
the first format considered.

As such, a format component is added to the font scoring heuristic.